requirements.txt
python-telegram-bot==21.5
openai==1.12.0
python-dotenv==1.0.0
aiohttp==3.9.1
telethon==1.34.0
Pillow==10.2.0